ikhal editor: bugfixes
ikhal does not exit on pressing 'esc' in event editor anymore, but
instead returns to the default view if the event has not been edited
yet, and does nothing otherwise (will be changed). Might have fixed a
bug in editing location and description properties, but I'm too lazy to
revert to the last commit just to check.