Various bug fixes.
Event creation issue:
  The dump function expects a dictionary to be passed in and then recursively
calls it self passing in the values in pulls out of the dictionary which led to
it crashing.  I added a check to make sure it is being passed a dictionary
otherwise it returns.  There is probably a better fix but I didn't completely
review what this function is used for/where it is called from.

Event upload issues - uploading to an Apple xserve iCal server.:
  When the calendar is created the prodid and version areset so it can be
properly parsed.
  When the even is created the DTSTAMP stamp
field is populated with the current datetime.  DTSTAMP is required but it is
only used to mark when the event was created which justified the use of
datetime.datetime.now().
  The time zone information at the beginning of DTSTART and DTEND were
causing 'Invalid calendar data' errors, changed aux so the timezone information
is no longer included.

iKhal:
  iKhal would triger ipdb.trace() whenever an event was saved.  Commented out
that line.