Changes to get timezone information working.
Most of this commit is adding the _create_timezone method to caldav which
creates the VTIMEZONE section of the ics.  Right now it uses the
default_timezone from the conf file but that can be changed easily.  The
_create_calendar function was created as well to prevent from cluttering up the
upload function.
Changed controllers to pass the confs along when there is going to be an upload.
Change to ui/__init__.py - toggle was not set anywhere, so if one of those
conditions wasn't met it would crash with toggle referenced before assigned.