Expand environment variables in config, fix sample config to match
The default config values for the sqlite path already makes use of the
XDG basedir spec.  However, user-supplied paths with environment
variables (either for the sqlite path or for calendar paths) were not
expanded before. This patch allows to use environment variables in these
config settings.

It also fixes the documentation so it mentions the XDG variables it
uses.