Use native strings for logging
The fix for #114 doesn't work if the item contains nonascii characters.
The ical variable seems to be a bytestring, and when formatting it into
a unicode string, Python wants to decode it with the ascii codec. Doing
so raises a UnicodeError (instead of UpdateFailed) and we still get an
error message with normal logging settings.