Add testcase
At some point there was a bug in khal where it would call _update_vevent
for everything in the vdir if something changed, because sqlite returned
None as etag for every item. I have no idea what fixed this bug, but
this is the testcase i wrote while trying to fix it.