Use deterministic UIDs on event from string tests.
Background: Previous tests use the UID that python2 will auto-generate when
`random` is seeded with `1`. Different versions of python return different
values, and we shouldn't rely the assumption that random can be made non-random
anyway.

Update the tests to set the UID of generated events to a pre-defined on, so that
tests can be written to match with these expected values.