refactoring in the backend
mostly in khalendar.khalendar and khalendar.backend.
If we trust the coverage (I don't :( ), this should be good to merge.

Long term backend.SQLiteDB and khalendar.khalendars API should be
cleaned up some more. The placement of `calendars` and if it's needed
and all is not really consistent, as naming of methods, properties and
arguments needs some work.