fix: handling of events without DTEND (#312)
When we find an event without a DTEND parameter it is treated as if it
were an all-day event lasting for the day its DTSTART is on. Those
start and end dates were saved to the database. After retrieving the
event from the database, the Event() type was inferred from the
(original) DTSTART property, which is not touched, but the transformed
start and end times were used. This lead to a discrepancy that lead to
a crash. Now, (transformed) start and end times are used to infer the
Event() type.

This should fix #312.