agenda: blank line before new day
This should improve legibility of the agenda. This style changed was
suggested several times in the issue tracker. If, however, this breaks
your emacs configuration (or you would like the old behaviour back for
other reasons), please voice your concerns, we can always introduce a
new config option.

This fixes #267