Don't expand events with REC.-ID but master event
Until now we expanded *all* events with a RRULE or RDATE, even if only
VEVENTs (with the same UID) with specified RECURRENCE-ID parameters were
present. In this case, the expansion is nonsensical, because we do not
know the start (and end) datetime of the master (but only of the
exceptions).

This fixes #323.