FIX: keep track of ANSI escape codes when wrapping.
As we keep using textwrap.wrap, we *could* split a line in the middle of
an escape sequence, but let's hope most people don't spray too many
color codes in their format directives.

Also, when the line length is measured, color codes are included, this
may lead to shorter lines than expected.

This fixes #569.