Refactor ikhal's DeleteDialog
Original did this to display warning errors if we do not completely
understand the event, which we do not currently need it for (see #532),
but we are going to use it soon for exciting messages such as "You are
not the organizer of this event, do you still want to edit it?" or "We
cannot reproduce the recurrence rule of this event, do you still want
to edit it?".

Backtracing can/should probably also moved in there, with some more
abstraction we might be able to make ExportDialog and SearchDialog use
it as well.