Refactor backend and introduce type hinting
Some logic gets moved out of the SQLite backend. Until now, the SQLiteDb
class accepted VEVENTs as strings, but returned our own Event instances.
Now, it accepts and returns strings. Some reshuffling of tests was
needed as well.

Also, this commit introduces PEP484 type hinting for all methods and
functions in khalendar.py and backend.py. This is to be expanded to
other parts.