[icalendar] Support rrules that ended in the past
Currently we were throwing an UnsupportedRecurrenceError when there
was an rrule that happened to have been finished (RRULE:UNTIL) in the
past.

This just returns an empty list of events when expanding instead.

This also adds freezegun (through pytest-freezegun) to the
khalendar_utils_test in order to be able to test this case, as most
tests were using old dates.

Signed-off-by: David Caro <david@dcaro.es>