until/every empty string should be mapped to None
Fix #999.

This also addresses a defect where editing a previously saved event
cause the 3 rrule properties (freq, until and interval) to have
incorrect default type (list instead of str), and freq value is forced
to lower case as the current code requires it but more important that is
how the values are presented in the interactive session.