Add support for processing and importing ics without UID
Files like this seem to be produced by all kinds of proprietary reservation
systems and such. They are clearly out of spec, but since they exist I'd
prefer to be able to import them without needing to manually add an UID
every time.

A sha256 checksum of the event is used as the UID to make it stable.