Do not evaluate etag when deleting instances
If multiple instances of an event are deleted one after the other,
the event will be updated each time and thus its etag will change dynamically
each time.
This will lead to a mismatch with the static etag recorded at the time
when the instances are queued for deletion.

Instead of overcomplicating things by re-adjusting the deletion queue
each time, do not pass the static etag so that it will not be
evaluated (delete_instance function for the collection, see
khalendar.py:229, i.e. do not trigger etag matching in khalendar.py:242).