Do not evaluate etag when deleting instances
If multiple instances of an event are deleted one after the other, the event
will be updated each time and thus its etag will change dynamically each time.
This will lead to a mismatch with the static etag recorded at the time when the
instances are queued for deletion.

Instead of re-adjusting the deletion queue each time, do not compare the
dynamic etag with the static etag. The etag for the event will be updated
anyway.