ikhal: fix deleting of multiple instances
When deleting multiple instances of the same recurring event, ikhal
would crash because of mismatching etags. This would happen, because we
delete one instance after the next, each time updating the etag. As we
check for matching etags when deleting (the one in the database and the
one stored in ikhal), this would crash ikhal. Matching the etag makes
sense, as we might overwrite freshly synced data otherwise. For now, the
server wins.