fix an issue with updating the eventlist after editing
After we edited an event, the event list would get updated between the
first and last date shown in calendar, not between the first and last
date shown in the event list. This would often lead to not updating the
right dates. This *should* only be an issue when the recurrence rules
change, but there also is an issue where the RecurrenceEditor spuriously
claims it has been edited (when it has not been edited).