fixes defaults for new event types
`default_event_duration` and `default_dayevent_duration` where mixed up. The
default for `default_event_duration` was 1 day and it was used as the
default lengths for new all-day events, now it is used for datetime
events and the default is 1 hour. The opposite was true for
`default_dayevent_duration`.