Drop support for Python 3.8
CI is failing for Python 3.8. It's not clear to me what's wrong (the
error doesn't make sense to me), and Python 3.8 is no longer supported,
so I doubt that reporting this upstream will lead anywhere.

Drop support for Python 3.8, which has reached its end of life and is no
longer supported by the ecosystem in general.